fix(types): Move maps.ts types to types lib
This allows me to set restrictions on the DetectionResult interface based on WasteClassesType

ref #34
