+++ apps/mull-ui/src/app/pages/waste-recognition-page/identified-waste-modal.tsx
@@ -1,10 +1,9 @@
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Dialog } from '@material-ui/core';
-import { DetectionResult } from '@mull/types';
+import { DetectionResult, wasteClassMap } from '@mull/types';
 import { WasteTypeSvgMap } from 'apps/mull-ui/src/constants';
 import React from 'react';
-import { classMap } from '../../services/maps';
 import './identified-waste-page.scss';
 
 export interface IdentifiedWasteModalProps {
@@ -20,7 +19,7 @@ export const IdentifiedWasteModal = ({
   open,
   setOpen,
 }: IdentifiedWasteModalProps) => {
-  const mapEntry = classMap[detectionResult?.class];
+  const mapEntry = wasteClassMap[detectionResult?.class];
 
   return (
     <Dialog
